+++ src/parser/mod.rs
@@ -248,6 +248,7 @@ mod tests {
             .into_iter()
             .collect(),
             levels: Default::default(),
+            multiline: false,
         }
     }